main, rules: Rename Jxxx to AppleJxxx
This makes it more clear that we're matching against Apple machines. The
kernel driver will be updated to use this naming scheme. For backwards
compatibility, we add a rule to rename J313 at runtime (the only enabled
model at this time).

This update must be released together with a matching asahi-audio update
(but the kernel can come later).

Signed-off-by: Hector Martin <marcan@marcan.st>