Use systemd for runtime/statedirectories and dynamic users
In systemd units, we can use RuntimeDirectory= and StateDirectory= to
create /run/speakersafetyd and /var/lib/speakersafetyd, and have it be
owned by the user running the service automatically.

This makes permission management less brittle than having the `make
install` phase or systemd-tmpfiles take care of applying ownership.

Some build sandboxes (like the NixOS one) don't propagate users defined
on the global system, but provide a almost-empty `/etc/passwd` and
`/etc/group`, so `install -o $user -g $group` cannot look up a user
that's not available in the sandbox.

In fact, setting up the global user is not required at all, we can
dynamically [allocate a user on-demand][0pointer-dynamic-users],
which also brings some more sandboxing and isolation.

Creating the `blackbox` subdirectory inside the state directory and
setting up more restrictive permissions is left to the process itself.
Theoretically, we could just be more restrictive on the state directory
itself, and use StateDirectoryMode=, though I wanted to keep the
existing behaviour for now.

Closes #25.

[0pointer-dynamic-users]: https://0pointer.net/blog/dynamic-users-with-systemd.html

Signed-off-by: Florian Klink <flokli@flokli.de>